Adjust showing default to the previous commit
In the previous commit having no default in the description was an empty
string so you can check if the value has content to add the default. Now
it was changed to "no" so we need to check if the value is not "no".

(cherry picked from commit b3a5b76ea299ad75919599979c8271fbfa4f0b86)