Use first contexts as alternative for the description groups in Help
This will not be the final solution but for release 9.1 it is still
helpfull as at least you have description groups for the help
functionality.

Later we could decide to adapt to add descriptions in line with other
images and/or consider for another solution.

(cherry picked from commit 5345ffadcc5981e71c42c845966694bbcecda6b6)